llmops/requirements.txt
ai-llmops-project==0.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
argcomplete==3.5.1
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
azure-ai-documentintelligence==1.0.0b4
azure-ai-ml==1.21.1
azure-common==1.1.28
azure-core==1.31.0
azure-core-tracing-opentelemetry==1.0.0b11
azure-cosmos==4.7.0
azure-identity==1.19.0
azure-mgmt-core==1.4.0
azure-monitor-opentelemetry==1.6.2
azure-monitor-opentelemetry-exporter==1.0.0b31
azure-search-documents==11.5.1
azure-storage-blob==12.23.1
azure-storage-file-datalake==12.17.0
azure-storage-file-share==12.19.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==43.0.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
docstring_parser==0.16
executing==2.1.0
fastapi==0.115.3
filelock==3.16.1
filetype==1.2.0
fixedint==0.1.6
Flask==3.0.3
Flask-Cors==4.0.2
flask-restx==1.3.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.21.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.27.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==24.3.1
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langchainhub==0.1.21
langsmith==0.1.136
MarkupSafe==3.0.2
marshmallow==3.22.0
matplotlib-inline==0.1.7
more-itertools==10.5.0
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.52.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-dbapi==0.48b0
opentelemetry-instrumentation-django==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-psycopg2==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-urllib==0.48b0
opentelemetry-instrumentation-urllib3==0.48b0
opentelemetry-instrumentation-wsgi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.2
portalocker==2.10.1
prompt_toolkit==3.0.47
promptflow==1.16.1
promptflow-azure==1.16.1
promptflow-core==1.16.1
promptflow-devkit==1.16.1
promptflow-tracing==1.16.1
proto-plus==1.25.0
protobuf==4.25.5
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.2
pydash==7.0.7
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.34
stack-data==0.6.3
starlette==0.41.1
strictyaml==1.7.3
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-requests==2.32.0.20240907
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
waitress==2.1.2
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.11.0
zipp==3.20.2
